final touches

+++ lib/models/customer.py
@@ -1,13 +1,13 @@
-from __init__ import CURSOR, CONN
+from models.__init__ import CURSOR, CONN
 
 class Customer:
     all = {}
 
     def __init__(self, id, name, email, phone):
-        self.id = id 
-        self.name = name
-        self.email = email
-        self.phone = phone
+        self._id = id 
+        self._name = name
+        self._email = email
+        self._phone = phone
     
     def __repr__(self):
         return f"<Customer {self.id}: {self.name}, {self.email}, {self.phone}>"
@@ -44,37 +44,59 @@ class Customer:
             self.phone = phone
         else:    
             raise ValueError("Phone must be a 10-digit integer.")
+        
+    @classmethod
+    def create_table(cls):
+        sql = """
+            CREATE TABLE IF NOT EXISTS customer (
+            id INTEGER PRIMARY KEY,
+            name TEXT ,
+            email TEXT,
+            phone INTEGER
+        )
+        """
+        CURSOR.execute(sql)
+        CONN.commit()
 
     @classmethod
-    def insert_customer(self):
+    def drop_table(cls):
+        sql = """
+            DROP TABLE IF EXISTS customer;
+        """
+        CURSOR.execute(sql)
+        CONN.commit()
+
+    @classmethod
+    def insert_customer(cls, name, email, phone):
         sql = """
             INSERT INTO customer (name, email, phone)
             VALUES (?, ?, ?)
-    """ 
+        """ 
 
-        CURSOR.execute(sql, (self.name, self.email, self.phone))
+        CURSOR.execute(sql, (name, email, phone))
         CONN.commit()
-
-        self.id = CURSOR.lastrowid
-        type(self).all[self.id] = self
 
     @classmethod
     def find_by_id(cls, id):
         sql = """
             SELECT *
-            FROM authors
+            FROM customer
             WHERE id = ?
         """
 
         row = CURSOR.execute(sql, (id,)).fetchone()
         return cls.instance_from_db(row) if row else None
 
-def delete_customer(customer_id):
-        sql = """
-        DELETE FROM customer 
-        WHERE id = ?
-        """
-    
-        CURSOR.execute(sql, (customer_id,))
-        CONN.commit()
+    def delete_customer(customer_id):
+        customer = Customer.find_by_id(customer_id)
+        if customer:
+            sql = """
+            DELETE FROM customer
+            WHERE id = ?
+            """
+            CURSOR.execute(sql, (customer_id,))
+            CONN.commit()
+            print(f"Customer with ID {customer_id} has been deleted.")
+        else:
+            print(f"Customer with ID {customer_id} does not exist.")
